+++ src/ocean/util/container/map/model/BucketElementGCAllocator.d
@@ -34,3 +34,22 @@ public class BucketElementGCAllocator(Bucket) : IBucketElementGCAllocator
         return new Bucket.Element;
     }
 }
+
+
+/*******************************************************************************
+
+    Creates an instance of BucketElementGCAllocator which is suitable for usage
+    with the Map type passed as a template parameter.
+
+    Template Params:
+        Map = the type to create the allocator according to
+
+    Returns:
+        an instance of type BucketElementGCAllocator class
+
+*******************************************************************************/
+
+public BucketElementGCAllocator!(Map.Bucket) instantiateAllocator ( Map ) ( )
+{
+    return new BucketElementGCAllocator!(Map.Bucket);
+}